weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "google-cloud-aiplatform>=1.132.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
]
